move search service to parent folder
