+++ angular-ui/src/app/browse/movie-list/movie-list.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MoviesService } from '../../shared/movies.service';
 import { Movie } from '../../shared/movie.model';
 import { DatePipe } from '@angular/common';
-import { SearchService } from '../search-by-text/search.service';
+import { SearchService } from '../search.service';
 
 @Component({
   selector: 'app-movie-list',
